src/requirements.txt
pydantic-ai
pydantic
tavily-python